qr_id_card_app/requirements.txt
Flask==2.3.2
gspread==5.8.0
google-auth==2.20.0
gunicorn==20.1.0
qrcode==7.4.2
Pillow==10.3.0
oauth2client==4.1.3